One component per file for about window.
