+++ app/js/stores/rpc/stats/credits.js
@@ -14,7 +14,11 @@ var CreditsRPCStore = Reflux.createStore({
     },
 
     getInitialState: function() {
-        return {};
+        if (this.data) {
+            return this.data;
+        } else {
+            return {};
+        }
     },
 
     hasData: function() {
